refactor: Registry 클래스, ControllerHandlerAdapter 클래스 패키지 이동
